Archive/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.24.0
google-cloud-aiplatform==1.38.1
python-multipart==0.0.7
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
google-genai==1.29.0
